Tests/LibWeb/Text/expected/hit_testing/overflow-hidden-wrapped-into-css-translate.txt

Transform parent, overflow child:
<DIV id="box">
Transform and overflow on same element:
<A id="target">
